+++ packages/compiler-cli/src/ngtsc/typecheck/extended/src/template_checks/invalid_banana_in_box/index.ts
@@ -8,8 +8,10 @@
 
 import {TmplAstBoundEvent, TmplAstNode, TmplAstRecursiveVisitor} from '@angular/compiler';
 import * as ts from 'typescript';
+
 import {ErrorCode} from '../../../../../diagnostics';
-import {TemplateCheck, TemplateContext, TemplateDiagnostic} from '../../../api';
+import {NgTemplateDiagnostic} from '../../../../api';
+import {TemplateCheck, TemplateContext} from '../../../api';
 
 /**
  * Ensures the two-way binding syntax is correct.
@@ -20,7 +22,7 @@ export class InvalidBananaInBoxCheck implements TemplateCheck<ErrorCode.INVALID_
   code: ErrorCode.INVALID_BANANA_IN_BOX = 8101;
 
   run(ctx: TemplateContext,
-      template: TmplAstNode[]): TemplateDiagnostic<ErrorCode.INVALID_BANANA_IN_BOX>[] {
+      template: TmplAstNode[]): NgTemplateDiagnostic<ErrorCode.INVALID_BANANA_IN_BOX>[] {
     const visitor = new BananaVisitor(ctx);
 
     return visitor.getDiagnostics(template);
@@ -28,7 +30,7 @@ export class InvalidBananaInBoxCheck implements TemplateCheck<ErrorCode.INVALID_
 }
 
 class BananaVisitor extends TmplAstRecursiveVisitor {
-  private diagnostics: ts.Diagnostic[] = [];
+  private diagnostics: NgTemplateDiagnostic<ErrorCode.INVALID_BANANA_IN_BOX>[] = [];
 
   constructor(public readonly ctx: TemplateContext) {
     super();
@@ -45,17 +47,16 @@ class BananaVisitor extends TmplAstRecursiveVisitor {
     if (name.startsWith('[') && name.endsWith(']')) {
       const boundSyntax = boundEvent.sourceSpan.toString();
       const expectedBoundSyntax = boundSyntax.replace(`(${name})`, `[(${name.slice(1, -1)})]`);
-      this.diagnostics.push(
-          this.ctx.templateTypeChecker.makeTemplateDiagnostic<ErrorCode.INVALID_BANANA_IN_BOX>(
-              this.ctx.component, boundEvent.sourceSpan, ts.DiagnosticCategory.Warning,
-              ErrorCode.INVALID_BANANA_IN_BOX,
-              `In the two-way binding syntax the parentheses should be inside the brackets, ex. '${
-                  expectedBoundSyntax}'. 
+      this.diagnostics.push(this.ctx.templateTypeChecker.makeTemplateDiagnostic(
+          this.ctx.component, boundEvent.sourceSpan, ts.DiagnosticCategory.Warning,
+          ErrorCode.INVALID_BANANA_IN_BOX,
+          `In the two-way binding syntax the parentheses should be inside the brackets, ex. '${
+              expectedBoundSyntax}'. 
                 Find more at https://angular.io/guide/two-way-binding`));
     }
   }
 
-  getDiagnostics(template: TmplAstNode[]): ts.Diagnostic[] {
+  getDiagnostics(template: TmplAstNode[]): NgTemplateDiagnostic<ErrorCode.INVALID_BANANA_IN_BOX>[] {
     this.diagnostics = [];
     for (const node of template) {
       node.visit(this);
